+++ .gitignore
@@ -1,0 +1,2 @@
+dist-newstyle
+cabal.project.local*
