+++ CHANGELOG.md
@@ -1,0 +1,5 @@
+# Revision history for MIPS-hs
+
+## 0.1.0.0 -- YYYY-mm-dd
+
+* First version. Released on an unsuspecting world.
