+++ MIPS-hs.cabal
@@ -1,0 +1,36 @@
+cabal-version:      2.4
+name:               MIPS-hs
+version:            0.1.0.0
+
+-- A short (one-line) description of the package.
+-- synopsis:
+
+-- A longer description of the package.
+-- description:
+
+-- A URL where users can report bugs.
+-- bug-reports:
+
+-- The license under which the package is released.
+-- license:
+author:             Kasdejong
+maintainer:         k.j2.de.jong@student.vu.nl
+
+-- A copyright notice.
+-- copyright:
+-- category:
+extra-source-files:
+    CHANGELOG.md
+    README.md
+
+executable MIPS-hs
+    main-is:          Main.hs
+
+    -- Modules included in this executable, other than Main.
+    other-modules: Mips.Alu, Mips.CPU, Mips.Memory, Mips.Operations, Mips.Util
+
+    -- LANGUAGE extensions used by modules in this package.
+    -- other-extensions:
+    build-depends:    base ^>=4.16.4.0, binary, containers
+    hs-source-dirs:   app
+    default-language: Haskell2010
