+++ app/Main.hs
@@ -1,0 +1,6 @@
+module Main where
+
+import Mips.CPU
+
+main :: IO ()
+main = putStrLn "Hello, Haskell!"
